Creates a new 'users-router.js' file inside of the 'users' folder to build a route for any GET requests to the '/api/users' endpoint; also moves the 'only()' and 'restricted()' middleware functions into the 'users-router.js' file as well
